requirements.txt
firebase-admin 
pyrebase4
yfinance
news-please
requests
beautifulsoup4
torch
transformers